fully configurable and motor control working

+++ device/lib/vehicule/includes/car.hpp
@@ -7,11 +7,11 @@ class Car {
 public:
 	Car(DCMotor front, DCMotor back);
 
-	void steerLeft(int16_t power);
-	void steerRight(int16_t power);
+	void steerLeft(int32_t power);
+	void steerRight(int32_t power);
 
-	void forward(int16_t power);
-	void backward(int16_t power);
+	void forward(int32_t power);
+	void backward(int32_t power);
 
 	void stopSteer();
 	void stopThrotle();
