+++ device/lib/vehicule/src/firmata_iot_control.cpp
@@ -4,18 +4,18 @@
 
 FirmataIotControl::FirmataIotControl(const std::string& port) : firmataClient_(port) {}
 
-void FirmataIotControl::setPinMode(uint8_t pin, uint8_t mode) {
+void FirmataIotControl::setPinMode(int32_t pin, int32_t mode) {
 	this->firmataClient_.setPinMode(pin, mode);
 }
 
-void FirmataIotControl::setPinHigh(uint8_t pin) {
+void FirmataIotControl::setPinHigh(int32_t pin) {
 	this->firmataClient_.setPinHigh(pin);
 }
 
-void FirmataIotControl::setPinLow(uint8_t pin) {
+void FirmataIotControl::setPinLow(int32_t pin) {
 	this->firmataClient_.setPinLow(pin);
 }
 
-void FirmataIotControl::setPWM(uint8_t pin, uint16_t value) {
+void FirmataIotControl::setPWM(int32_t pin, int32_t value) {
 	this->firmataClient_.setPWM(pin, value);
 }
